django_ambassador/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
async-timeout==4.0.3; python_version >= '3.7'
django==5.0.2; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
django-mysql==4.12.0; python_version >= '3.8'
django-redis==5.4.0; python_version >= '3.6'
djangorestframework==3.14.0; python_version >= '3.6'
faker==23.3.0; python_version >= '3.8'
mysqlclient==2.2.4; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
redis==5.0.2; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
tzdata==2024.1; sys_platform == 'win32'
